+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 3.0.0-dev
 - BC break: moved PHP attributes to namespace MyTester\Attributes
 - automated tests runner now uses Composer's autoloader, package nette/robot-loader is optional
 - BC break: marked Job as final
+- automated tests runner's output can be colorized
 
 Version 2.1.0
 - allowed customization of test suite creation in automated tests runner